Python/Weather_App/requirements.txt
Pillow
requests
staticmap